PHPStan Level 2 (#89)
* initialize phpstan level 2

* improve method naming

* improve method naming

* add missing methods to user interface

+++ src/MembersBundle/Manager/UserManager.php
@@ -54,6 +54,10 @@ class UserManager implements UserManagerInterface
      */
     public function deleteUser(UserInterface $user)
     {
+        if (!$user instanceof DataObject\Concrete) {
+            return false;
+        }
+
         return $user->delete();
     }
